+++ backend/src/main/java/kakaobootcamp/backend/domains/autoTrade/AutoTradeService.java
@@ -15,7 +15,7 @@ import kakaobootcamp.backend.domains.broker.service.BrokerService;
 import kakaobootcamp.backend.domains.member.MemberService;
 import kakaobootcamp.backend.domains.member.domain.AutoTradeState;
 import kakaobootcamp.backend.domains.member.domain.Member;
-import kakaobootcamp.backend.domains.stock.StockService;
+import kakaobootcamp.backend.domains.stock.service.StockService;
 import kakaobootcamp.backend.domains.stock.dto.StockDTO.GetStockBalanceResponse;
 import kakaobootcamp.backend.domains.stock.dto.StockDTO.GetStockBalanceResponse.Output1;
 import kakaobootcamp.backend.domains.stock.dto.StockDTO.OrderStockRequest;
